Making a skip line function
The code worked on “raw” text, but if I fed it any xml, the tags threw it for a loop. Since my usual workflow is to mark up the tokens first and then dive into apodoses, I had to fix this. I started yesterday and it wound up being a convoluted mess. Falling asleep last night I realized a different solution, and during a break this morning, solved it. So back to working again.