Added visual breaks between tasks
I realized that it was easy to get lost in the wall of text, and I’d sometimes give commands for aposis-modification when I should have been giving commands for sense-modification and vice-versa. So now there’s a big visual break between each. Two lines of a red question mark ❓ emoji for the apodosis task and four lines of w=white speech bubble 💬 emoji for the sense task. This should help keep you more on task.